backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
alembic
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic
psycopg2-binary
python-dotenv
httpx
email-validator
python-multipart
sendgrid